ajoute des messages à chaque statut de refresh et déplace le code du refresh dans le package refresh
