+++ src/core/import.go
@@ -5,14 +5,10 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v4/pgxpool"
-	"github.com/signaux-faibles/datapi/src/refresh"
 	"io"
 	"log"
 	"math"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -880,61 +876,4 @@ func queueScoreToBatch(s scoreFile, batch *pgx.Batch) {
 		s.AlertPreRedressements,
 		s.Redressements,
 	)
-}
-
-func ExecRefreshScript(ctx context.Context, db *pgxpool.Pool, scriptPath string) uuid.UUID {
-	current := refresh.New(uuid.New())
-	sql, err := os.ReadFile(scriptPath)
-	if err != nil {
-		current.Fail(err.Error())
-		return current.Id
-	}
-	if len(sql) <= 0 {
-		current.Fail("le script sql est vide")
-		return current.Id
-	}
-	go executeRefresh(ctx, db, string(sql), current)
-	return current.Id
-}
-
-func executeRefresh(ctx context.Context, db *pgxpool.Pool, sql string, refresh *refresh.Refresh) {
-	tx, err := db.Begin(ctx)
-	if err != nil {
-		log.Fatalf("Erreur à l'ouverture de la transaction pour le refresh des vues : %s", err.Error())
-	}
-	for _, current := range strings.Split(sql, ";\n") {
-		log.Printf("Refresh - %s - Exécute la requête : '%s'", refresh, current)
-		refresh.Run(current)
-		_, err = tx.Exec(ctx, current)
-		if err != nil {
-			refresh.Fail(err.Error())
-			err := tx.Rollback(ctx)
-			if err != nil {
-				log.Fatalf("Erreur lors du rollback de transaction. Cause : %s", err.Error())
-			}
-			log.Fatalf("Erreur lors de l'exécution de la requête '%s'. Cause : %s", current, err.Error())
-		}
-	}
-	err = tx.Commit(ctx)
-	if err != nil {
-		refresh.Fail(err.Error())
-		log.Fatalf("Erreur lors du commit de transaction : %s", err.Error())
-	}
-	refresh.Finish()
-}
-
-func FetchRefreshStatus(uuid uuid.UUID) (refresh.Refresh, error) {
-	//uuid, err := uuid.Parse(refreshID)
-	//if err != nil {
-	//	return refresh.Empty, err
-	//}
-	return refresh.Fetch(uuid)
-}
-
-func FetchLastRefreshState() refresh.Refresh {
-	return refresh.FetchLast()
-}
-
-func FetchRefreshWithState(status refresh.Status) []refresh.Refresh {
-	return refresh.FetchRefreshWithState(status)
 }
